refactor: DAH-4291 move logging + browser-snapshot helpers out of the services
Two reorganizations from review.

Ruby: format_duration, scrubbed_referrer, and deadline_terms move from
InviteToController into InviteToEventLogging. All three exist only to turn raw
request values into loggable fields, so they belong beside log_invite_event.
scrubbed_referrer needs `request`, which the concern already uses for
request_id, so nothing had to be threaded through. Specs are unchanged - they
reach these via `controller.send`, and included-module methods resolve the same.

TypeScript: BrowserSnapshot / snapshotBrowser move out of inviteToApiService,
which is for API calls and never used them, into util/browserSnapshotUtil.ts
alongside the other *Util modules.

A separate module rather than folding them into the hook, for two reasons:
BrowserSnapshot types HumanVerifiedRecord.browser, so keeping it beside the
payload avoids the hook <-> api circular import that moving it would create;
and module-scoped it stays mockable, so the hook test keeps its fixture instead
of asserting against whatever jsdom happens to report.

Tests moved with the code: snapshotBrowser's cases now live in
__tests__/util/browserSnapshotUtil.test.ts, the hook test mocks the new module,
and invite-to-apply.test.tsx drops its stale mock entry.
