+++ app/javascript/__tests__/pages/invite-to-apply.test.tsx
@@ -16,7 +16,6 @@ jest.mock("../../api/inviteToApiService", () => ({
   recordResponse: jest.fn(),
   logHumanVerifiedClick: jest.fn(),
   beaconHumanVerifiedClick: jest.fn(),
-  snapshotBrowser: jest.fn(() => ({})),
 }))
 jest.mock("../../hooks/useFeatureFlag", () => ({
   useFeatureFlag: () => ({
